changes to accomodate reorganization of Policy classes to rl/policy.py
