+++ rl/problems/Final-Winter2021/windy_grid_outline.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 from rl.distribution import Categorical, Choose
 from rl.markov_process import NonTerminal
 from rl.markov_decision_process import FiniteMarkovDecisionProcess
-from rl.markov_decision_process import FiniteDeterministicPolicy
+from rl.policy import FiniteDeterministicPolicy
 from rl.dynamic_programming import value_iteration_result, V
 from operator import itemgetter
 
@@ -260,7 +260,7 @@ class WindyGrid:
                                         for j in range(self.columns)))
         print()
         pol_full_dict = {
-            **{s: possible_moves[policy.deterministic_policy_map[NonTerminal(s)]]
+            **{s: possible_moves[policy.action_for[s]]
                for s in self.get_all_nt_states()},
             **{s: 'T' for s in self.terminals},
             **{s: 'X' for s in self.blocks}
