+++ rl/problems/Midterm-Winter2021/career_optimization.py
@@ -1,5 +1,4 @@
 from typing import Tuple, Mapping, Dict, Sequence, Iterable
-from rl.markov_process import NonTerminal
 from rl.markov_decision_process import FiniteMarkovDecisionProcess
 from rl.dynamic_programming import value_iteration_result
 from rl.distribution import Categorical
@@ -75,7 +74,7 @@ if __name__ == '__main__':
     _, opt_det_policy = value_iteration_result(co, gamma=gamma)
     wages: Iterable[int] = range(1, co.wage_cap + 1)
     opt_actions: Mapping[int, Tuple[int, int]] = \
-        {w: opt_det_policy.deterministic_policy_map[NonTerminal(w)]
+        {w: opt_det_policy.action_for[w]
          for w in wages}
     searching: Sequence[int] = [s for _, (s, _) in opt_actions.items()]
     learning: Sequence[int] = [l for _, (_, l) in opt_actions.items()]
