+++ rl/problems/Midterm-Winter2021/grid_maze.py
@@ -1,5 +1,3 @@
-from typing import Tuple, Mapping
-
 SPACE = 'SPACE'
 BLOCK = 'BLOCK'
 GOAL = 'GOAL'
